+++ src/api/person.js
@@ -1,17 +1,13 @@
-import { Request } from '../utils/request';
-
-const API_URL = 'http://localhost:3000';
+import requestFromAPI from '../utils/request';
 
 export default class Person {
   constructor(username, id) {
     this.id = id;
     this.username = username;
-
-    this.request = new Request(API_URL);
   }
 
   get() {
-    if (this.username !== null) return this.request.get(`person/get?username=${this.username}`);
-    return this.request.get(`person/get?id=${this.id}`);
+    if (this.username !== null) return requestFromAPI(`person/get?username=${this.username}`);
+    return requestFromAPI(`person/get?id=${this.id}`);
   }
 }